Restore compatibility with Visual C++ 2005.

+++ ExtLibs/pstdint.h
@@ -377,7 +377,9 @@
 #endif
 #ifndef uint32_t
 #if (ULONG_MAX == UINT32_MAX) || defined (S_SPLINT_S)
-  typedef unsigned long uint32_t;
+  // CF: Changed from `unsigned long` to `unsigned int`, or else Visual C++ 2005
+  // CF: complains that operator << in cf::Network::OutStreamT is ambiguous.
+  typedef unsigned int uint32_t;
 # define UINT32_C(v) v ## UL
 # ifndef PRINTF_INT32_MODIFIER
 #  define PRINTF_INT32_MODIFIER "l"
@@ -407,7 +409,9 @@
 #endif
 #ifndef int32_t
 #if (LONG_MAX == INT32_MAX) || defined (S_SPLINT_S)
-  typedef signed long int32_t;
+  // CF: Changed from `signed long` to `signed int`, or else Visual C++ 2005
+  // CF: complains that operator << in cf::Network::OutStreamT is ambiguous.
+  typedef signed int int32_t;
 # define INT32_C(v) v ## L
 # ifndef PRINTF_INT32_MODIFIER
 #  define PRINTF_INT32_MODIFIER "l"
